Rename transactionChart methods
Make ContextProperties more generic

Replace Spans with Waterfall component

Add flyout

Add span and transaction flyouts

Fix stacktraces

Add charts and distribution to flyout

Fix issue with ownFocus

Fix getPropertyTabNames
