+++ x-pack/plugins/apm/public/components/app/TransactionDetails/Trace/Waterfall/WaterfallItem.js
@@ -1,0 +1,86 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import React from 'react';
+import styled from 'styled-components';
+import PropTypes from 'prop-types';
+
+import {
+  unit,
+  units,
+  colors,
+  px,
+  fontFamilyCode,
+  fontSizes
+} from '../../../../../style/variables';
+
+const ItemBar = styled.div`
+  position: relative;
+  height: ${unit}px;
+`;
+const ItemLabel = styled.div`
+  white-space: nowrap;
+  position: relative;
+  direction: rtl;
+  text-align: left;
+  margin: ${px(units.quarter)} 0 0;
+  font-family: ${fontFamilyCode};
+  font-size: ${fontSizes.small};
+`;
+
+const Container = styled.div`
+  position: relative;
+  display: block;
+  user-select: none;
+  padding: ${px(units.half)} ${props => px(props.timelineMargins.right)}
+    ${px(units.eighth)} ${props => px(props.timelineMargins.left)};
+  border-top: 1px solid ${colors.gray4};
+  background-color: ${props => (props.isSelected ? colors.gray5 : 'initial')};
+  cursor: pointer;
+  &:hover {
+    background-color: ${colors.gray5};
+  }
+`;
+
+function WaterfallItem({
+  timelineMargins,
+  totalDuration,
+  waterfallItem,
+  color,
+  isSelected,
+  onClick
+}) {
+  const width = (waterfallItem.duration / totalDuration) * 100;
+  const left = (waterfallItem.offset / totalDuration) * 100;
+
+  return (
+    <Container
+      onClick={() => onClick(waterfallItem)}
+      timelineMargins={timelineMargins}
+      isSelected={isSelected}
+    >
+      <ItemBar
+        style={{
+          left: `${left}%`,
+          width: `${width}%`,
+          backgroundColor: color
+        }}
+      />
+      <ItemLabel style={{ left: `${left}%`, width: `${100 - left}%` }}>
+        &lrm;
+        {waterfallItem.name}
+        &lrm;
+      </ItemLabel>
+    </Container>
+  );
+}
+
+WaterfallItem.propTypes = {
+  location: PropTypes.object.isRequired,
+  totalDuration: PropTypes.number.isRequired
+};
+
+export default WaterfallItem;
