--- x-pack/plugins/apm/public/components/app/TransactionDetails/Transaction/Spans/index.js
@@ -1,22 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { connect } from 'react-redux';
-import Spans from './view';
-import { getUrlParams } from '../../../../../store/urlParams';
-
-function mapStateToProps(state = {}) {
-  return {
-    urlParams: getUrlParams(state),
-    location: state.location
-  };
-}
-
-const mapDispatchToProps = {};
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Spans);
