--- x-pack/plugins/apm/public/components/app/TransactionDetails/Transaction/index.js
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { connect } from 'react-redux';
-import Transaction from './view';
-
-function mapStateToProps(state = {}) {
-  return {
-    location: state.location
-  };
-}
-
-const mapDispatchToProps = {};
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Transaction);
