+++ x-pack/plugins/apm/public/components/app/TransactionDetails/view.js
@@ -7,13 +7,13 @@
 import React from 'react';
 import { EuiSpacer } from '@elastic/eui';
 import { HeaderLarge } from '../../shared/UIComponents';
-import Transaction from './Transaction';
 import Distribution from './Distribution';
-import { DetailsChartsRequest } from '../../../store/reactReduxRequest/detailsCharts';
-import Charts from '../../shared/charts/TransactionCharts';
 import { TransactionDistributionRequest } from '../../../store/reactReduxRequest/transactionDistribution';
-import { TransactionDetailsRequest } from '../../../store/reactReduxRequest/transactionDetails';
+import TransactionCharts from '../../shared/charts/TransactionCharts';
+import { TransactionDetailsChartsRequest } from '../../../store/reactReduxRequest/transactionDetailsCharts';
 import { KueryBar } from '../../shared/KueryBar';
+import mockTraceData from './mockTraceData.json';
+import Trace from './Trace';
 
 function TransactionDetails({ urlParams, location }) {
   return (
@@ -24,10 +24,14 @@ function TransactionDetails({ urlParams, location }) {
 
       <EuiSpacer size="s" />
 
-      <DetailsChartsRequest
+      <TransactionDetailsChartsRequest
         urlParams={urlParams}
         render={({ data }) => (
-          <Charts charts={data} urlParams={urlParams} location={location} />
+          <TransactionCharts
+            charts={data}
+            urlParams={urlParams}
+            location={location}
+          />
         )}
       />
 
@@ -38,11 +42,12 @@ function TransactionDetails({ urlParams, location }) {
         )}
       />
 
-      <TransactionDetailsRequest
+      <Trace
+        location={location}
+        services={mockTraceData.services}
+        trace={mockTraceData.trace}
         urlParams={urlParams}
-        render={({ data }) => (
-          <Transaction transaction={data} urlParams={urlParams} />
-        )}
+        waterfallItems={mockTraceData.waterfallItems}
       />
     </div>
   );
