+++ x-pack/plugins/apm/public/components/shared/charts/Timeline/TimelineAxis.js
@@ -18,7 +18,7 @@ import { getTimeFormatter } from '../../../../utils/formatters';
 const getXAxisTickValues = (tickValues, xMax) =>
   _.last(tickValues) * 1.05 > xMax ? tickValues.slice(0, -1) : tickValues;
 
-function TimelineAxis({ header, plotValues, agentMarks }) {
+function TimelineAxis({ plotValues, agentMarks }) {
   const { margins, tickValues, width, xDomain, xMax, xScale } = plotValues;
   const tickFormat = getTimeFormatter(xMax);
   const xAxisTickValues = getXAxisTickValues(tickValues, xMax);
@@ -38,13 +38,12 @@ function TimelineAxis({ header, plotValues, agentMarks }) {
               ...style
             }}
           >
-            {header}
             <XYPlot
               dontCheckIfEmpty
               width={width}
-              height={40}
+              height={margins.top}
               margin={{
-                top: 40,
+                top: margins.top,
                 left: margins.left,
                 right: margins.right
               }}
